+++ engine/launcher/src/main/scala/org/enso/launcher/cli/InternalOpts.scala
@@ -6,9 +6,9 @@ import cats.implicits._
 import nl.gn0s1s.bump.SemVer
 import org.enso.cli.arguments.Opts
 import org.enso.cli.arguments.Opts.implicits._
-import org.enso.distribution.OS
-import org.enso.runtimeversionmanager.{CurrentVersion, FileSystem}
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
+import org.enso.distribution.{FileSystem, OS}
+import org.enso.runtimeversionmanager.CurrentVersion
+import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.cli.Arguments._
 import org.enso.launcher.distribution.LauncherEnvironment
 import org.enso.launcher.upgrade.LauncherUpgrader
