+++ engine/launcher/src/main/scala/org/enso/launcher/upgrade/LauncherUpgrader.scala
@@ -4,15 +4,15 @@ import java.nio.file.{Files, Path}
 import com.typesafe.scalalogging.Logger
 import nl.gn0s1s.bump.SemVer
 import org.enso.cli.CLIOutput
-import org.enso.distribution.OS
+import org.enso.distribution.{FileSystem, OS}
 import org.enso.distribution.locking.{
   LockType,
   LockUserInterface,
   Resource,
   ResourceManager
 }
-import org.enso.runtimeversionmanager.{CurrentVersion, FileSystem}
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
+import org.enso.runtimeversionmanager.CurrentVersion
+import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.archive.Archive
 import org.enso.runtimeversionmanager.components.UpgradeRequiredError
 import org.enso.runtimeversionmanager.distribution.DistributionManager
