+++ engine/launcher/src/test/scala/org/enso/launcher/components/LauncherRunnerSpec.scala
@@ -4,7 +4,7 @@ import java.nio.file.{Files, Path}
 import java.util.UUID
 import akka.http.scaladsl.model.Uri
 import nl.gn0s1s.bump.SemVer
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
+import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.config.GlobalConfigurationManager
 import org.enso.runtimeversionmanager.runner._
 import org.enso.runtimeversionmanager.test.RuntimeVersionManagerTest
