+++ lib/scala/runtime-version-manager-test/src/test/scala/org/enso/runtimeversionmanager/components/RuntimeVersionManagerSpec.scala
@@ -2,15 +2,15 @@ package org.enso.runtimeversionmanager.components
 
 import java.nio.file.{Files, Path}
 import nl.gn0s1s.bump.SemVer
-import org.enso.distribution.OS
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
+import org.enso.distribution.{FileSystem, OS}
+import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.config.GlobalConfigurationManager
 import org.enso.runtimeversionmanager.releases.ReleaseNotFound
 import org.enso.runtimeversionmanager.test.{
   RuntimeVersionManagerTest,
   TestRuntimeVersionManagementUserInterface
 }
-import org.enso.runtimeversionmanager.{components, FileSystem}
+import org.enso.runtimeversionmanager.components
 import org.enso.testkit.OsSpec
 
 class RuntimeVersionManagerSpec extends RuntimeVersionManagerTest with OsSpec {
