+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/components/RuntimeVersionManager.scala
@@ -3,10 +3,10 @@ package org.enso.runtimeversionmanager.components
 import java.nio.file.{Files, Path, StandardOpenOption}
 import com.typesafe.scalalogging.Logger
 import nl.gn0s1s.bump.SemVer
-import org.enso.distribution.OS
+import org.enso.distribution.{FileSystem, OS}
 import org.enso.distribution.locking.{LockType, ResourceManager}
-import org.enso.runtimeversionmanager.{CurrentVersion, FileSystem}
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
+import org.enso.runtimeversionmanager.CurrentVersion
+import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.archive.Archive
 import org.enso.runtimeversionmanager.distribution.{
   DistributionManager,
