+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/distribution/DistributionManager.scala
@@ -2,9 +2,8 @@ package org.enso.runtimeversionmanager.distribution
 
 import java.nio.file.{Files, Path}
 import com.typesafe.scalalogging.Logger
-import org.enso.distribution.OS
-import org.enso.runtimeversionmanager.FileSystem.PathSyntax
-import org.enso.runtimeversionmanager.{Environment, FileSystem}
+import org.enso.distribution.{Environment, FileSystem, OS}
+import org.enso.distribution.FileSystem.PathSyntax
 
 import scala.util.Try
 import scala.util.control.NonFatal
